solution for  suspension system for one wheel of an old-fashioned  pickup truck is in the pdf
The suspension system for one wheel of an old-fashioned 
the pickup truck is illustrated in figure 2. The mass of the vehicle is m1 and the mass of the wheel is m2. The suspension spring has a spring constant k1 and the tire had the spring constant k2. The damping constant of the shock absorber is b. Obtain the mathematical model which represents the vehicle response to a bump in the road and simulate it for 100 sec. Create a dummy input signal to represent the bumpy road using 
MATLAB program.